Implement the `display_field()` field type method
The method is used for default field value filtering for the current field
type. Since BP 2.4 has added the `$field_id` parameter, now we can use the
fields' context to determine which relationship is defined, and
consequently what the proper filtered value should be. Before, this was
not possible to find out. Filter logic remains in the main plugin class.

Additionally, this solves a lot of special cases where by-type filtering
is assumed. This makes the filtering of the `bp_get_member_profile_data()`
values obsolete.

Also, the filtering of field data used in the template tag for data
display, is filtered separately, so no additional filters (like object
linking) are applied.